requirements.txt
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
billiard==4.2.0; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
django==4.2; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
kombu==5.3.7; python_version >= '3.8'
pillow==10.4.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.11'
tzdata==2024.1; python_version >= '2'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
billiard==4.2.0; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
django==4.2; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
kombu==5.3.7; python_version >= '3.8'
pillow==10.4.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.11'
tzdata==2024.1; python_version >= '2'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
